fix: deploy with hubspot api v3
Hubspot’s v2 api was not returning all pages, even though we were not hitting the limit.  This was causing some pages to be re-created even though they already existed.  Hubspot was appending an integer for a while then began appending a timestamp to avoid overwriting the existing page.  Looking at the timestamps of the published pages, this has been a problem since December 2020, and 80 copies of each page have been mistakenly published.

I tested the code on the `learn-docker` pages and it successfully updated the files.  Before publishing the entire site, I’m going to let somebody else take a look and run the code.

There was some previous work done to say how many pages need to be deleted.  I updated the code to log the slug of each duplicate/orphaned page, which shows the pages that end in timestamps (Hubspot duplicate prevention).  We will need to make another PR to delete the orphaned pages.

Improves debuggability of some code, as well.