Add a starter outline for the new React workshop (#447)
The current content on Bitovi.com is under `src/react/`, so as to not mess up the current Academy content, this adds a new `src/react-vite/` folder for the new workshop. When we’re done with building this workshop, we can move it under the [https://www.bitovi.com/academy/learn-react.html] URL.

This includes pages for all the sections we’ll cover. Each page follows a general structure:

- Overview
- Objective 1
	- Key concepts
		- Concept 1
		- Concept 2
	- Requirements
	- Setup
	- Verify
	- Solution
- Next steps

This is similar to the structure used on other Academy pages, but not the same; we identified some confusion with the other structure and we’re trying new words to improve it for the future.

In this commit, you’ll also find some diffs to point out what each section needs to teach in order for participants to complete the exercise. The goal here was to show the Place My Order app being built in parts so we can focus on exactly what we need to teach in each section.

This is just a rough draft and needs to be refined as we figure out how we want to teach each concept.