Move the Requirements to be before the Solution
After explaining the Overview/Objective/Concepts/Requirements/Setup/Verify/Solution structure to @rlmcneary2 and @rjspencer, we realized that it would make more sense for the “requirements” to be closer to the solution so folks know exactly what they’re trying to accomplish and have a reference while they’re coding a solution.

As part of this, we’re renaming Requirements to Exercise so it also aligns with how we structured the slides.