Refactor the order of Props and Routing
In addition to switching the Routing and Props pages (so Props now comes before Routing), this makes significant changes to the exercises.

- The “Intro to JSX” page now focuses on building the `RestaurantList` first (instead of the homepage)
- The “Components” page then continues by splitting the `RestaurantList` into its own component (from `App`)
- The “Props” page then introduces the restaurant `ListItem` as its own component (to introduce passing and using props)
- The “Routing” page now includes creating the `Home` component, then creating the routes and navigation

I believe that all of the exercise problem & solution folders have been updated and the diffs on the pages are correct, but the content on the pages needs to be updated to reflect the new structure.