Rearrange the React pages
Previous structure:

- Managing State
- Handling Side Effects
- Handling User Inputs
- Fetching Data
- Testing

New structure:

- Managing State
- Making HTTP Requests (was Fetching Data)
- Creating Nested Routes (was Handling Side Effects)
- Handling User Inputs
- Testing

Making HTTP Requests was moved up so we can actually fetch state, city, and restaurant data sooner.

Handling Side Effects was removed because the ideas for that section are now covered elsewhere; Creating Nested Routes was added so we have a bridge section for creating the Restaurant Detail and Order pages.