Rearrange most of the TypeScript files
This moves most of the files into folders so the structure is easier to navigate.

The other files will be moved in other PRs.

You’ll notice that this does not change the numbers in the file names in the folders. We can remove those numbers as we make changes to the individual pages.