Create a separate “Setting Up Your Environment” page
This takes the first exercise that was on the “Why TypeScript” page and moves it to a brand new, more thorough “Setting Up Your Environment” page.

As part of moving this, note that we will not use the [bitovi/learn-typescript](https://github.com/bitovi/learn-typescript) repo. Instead, we have learners create their own project locally and create the files as they follow along.