Move “Exporting and Importing” out of “Configuration” (#612)
This creates a separate page for the “Exporting and Importing” exercise that was part of the “Configuration” page. They seemed to have nothing to do with each other, so I think this makes more sense.

Note how the tests are done for the exercise. We instruct learners to add use this for the `test` script in their `package.json`:

```
node --require ts-node/register --test src/**/**.test.**
```

This will look for any `.test.` file and run it with [Node’s test runner](https://nodejs.org/api/test.html). We do not need Mocha or any other dependency.