Expand the “Setting Up Your Environment” page (#688)
This adds a lot more detail to the page about setting up macOS and Windows for React Native development. We will expand this content more as we have more people try to set up their environments before the training.

I debated about how the page should be organized. Even though the objectives are the same for macOS and Windows, I decided to keep the instructions separate so it’s easier to go through them on the page.

During the actual training (and in the slides), I imagine that we will go objective by objective, and attendees will need to stay in the section that’s right for their operating system.