Fix more exercise solution tests
In Navigation Params, the `<StateList />)` needed to be wrapped in a `<NavigationContainer>`.

In HTTP Requests, I made some bigger changes:

- Fixed the `package.json` and added a matching `package-lock.json` file.
- Since the `10-` problem & solution folders looked like they would be the last objective, I renamed them to `04-`.
- Applied some of @christopherjbaker’s changes from this PR into our code: https://github.com/bitovi/react-native-pmo/pull/49

For User Input, I also applied some of Christopher’s refactoring changes to make the tests pass.

There is more work to be done on the exercise problems & solutions, but this at least gets the solutions up to Async Storage passing.