Make the exercise test command more resilient to failures
I sometimes run into issues with the test script where it is unable to delete the `node_modules` folder after running an exercise solution’s test.

Since this step is not critical to whether the tests have passed or not, these errors will now be caught and logged, but ignored so the script can keep running.

I’d like to look into why I’m seeing these failures, but in the meantime, I think this improves the script for anyone else who might run into a similar issue.