Write the content for “Adding Offline Support”
The exercise code needs to be refactored, so the content does the best it can with the code as-is.

The content takes a departure from our normal approach by explaining the code that will be given in the Setup steps because there is a lot to explain about syncing, so it seems like using our code as an example will be the most useful way to explain the concepts.

Closes https://bitovi.atlassian.net/browse/LD-310