Refactor the exercises for “Adding Offline Support” (#813)
- Refactor the code to work correctly.
- Don’t return the server response from `useFavorites`.
- Do return the `error` and `isPending` state from `useFavorites` so the UI can show these values.
- Show an error if there’s a problem with signing in or out.
- Add a “Sign in to favorite this restaurant” button when the user is signed out.
- Move the interfaces to be in their own file.

The goal of this PR is to fix the solution code, so there will be a follow-up PR to make the problem folders, diffs, etc. correct.

Also, I didn’t completely overhaul the names of functions/types/etc., so please feel free to point out specific changes you’d like to see.